switch the doc of the tree to doc.rs
